move storeContracts into task utils
